Add RPC client plugin
