+++ packages/blitz-rpc/src/global.ts
@@ -1,0 +1,6 @@
+import {QueryClient} from "react-query"
+
+declare global {
+  var queryClient: QueryClient
+  var __BLITZ_SUSPENSE_ENABLED: boolean
+}
